requirements.txt
numpy
pandas
matplotlib
sklearn
mojimoji
mecab-python3
janome
spacy
SudachiPy
nagisa
lightgbm
line_profiler
memory_profiler
psutil
umap-learn
numba
pudb
gensim
flake8
sentencepiece
